% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/mnmc.R
\name{mnmc}
\alias{mnmc}
\title{Estimate Multivariate Normal Mean Coding}
\usage{
mnmc(X, W = NULL, Delta0 = NULL, ConvCriteria = 2, tol = 0.01)
}
\arguments{
\item{X}{ordinal predictor matrix, each row is an observation and each column a predictor variable.}

\item{W}{continuous predictor matrix, each row is an observation and each column a predictor variable.}

\item{Delta0}{initial covariance matrix}

\item{ConvCriteria}{convergence criteria. If =1 we take distance}

\item{tol}{convergence tolerance}
}
\value{
a matrix M with first moment estimates for the latent variables, given the observed Parameter Estimates: Delta, Theta; data.
}
\description{
This function estimates the conditional expectation of ordinal variables X, i.e. M=E(Z|X). Here we follow the first step algorithm of Forzani et.al (2018).
}
